refactor: Add full-page property to form for out of the box VR support
