+++ pages/form/permutations.page.tsx
@@ -62,6 +62,7 @@ const permutations = createPermutations<FormProps>([
         <Button variant="primary">Submit</Button>
       </SpaceBetween>,
     ],
+    fullPage: [false, true],
   },
 ]);
 
@@ -92,6 +93,7 @@ const secondaryActionsPermutations = createPermutations<FormProps>([
         <Button>Save and close</Button>
       </SpaceBetween>,
     ],
+    fullPage: [false],
   },
 ]);
